[REFACTOR] Replaced CarProcessing class from domain package to controller package
