--- src/test/java/racingcar/main.java
@@ -1,4 +1,0 @@
-package racingcar;
-
-public class main {
-}
